requirements.txt
pyinstaller
pyyaml